Remove trimming option. Provide way to use alternate artwork.

+++ youtube_songs/process_mp4.py
@@ -3,7 +3,6 @@ For all the .mp4 files in a given directory, do the following:
 
 - Extract the audio out into a .m4a file.
 - Adjust gain.
-- Optionally trim running time.
 - Add metadata (title, artist, album, artwork).
 
 Usage: python process_mp4_files.py /path/to/mp4/files/
@@ -12,14 +11,14 @@ Usage: python process_mp4_files.py /path/to/mp4/files/
 
 Each input file should use the following format for their filename:
 
-ALBUM  TITLE  ARTIST  [START-END]
+ALBUM  TITLE  ARTIST
 
 Note that there are two spaces between each element.
 
 ---
 
 In order to run this script, you must have ffmpeg, aacgain, and AtomicParsley
-installed. 
+installed.
 
 On Ubuntu:
 
@@ -50,46 +49,36 @@ import os.path as op
 import subprocess
 import shutil
 import tempfile
+from pathlib2 import Path
 
 
 class Song:
-    def __init__(self, mp4_path):
+    def __init__(self, mp4_file):
         self.input_file = mp4_file
-        name = op.basename(mp4_file).rsplit('.', 1)[0]
-        parts = name.split('  ')
-        self.album, self.title, self.artist = parts[:3]
-        
-        dirname = op.dirname(mp4_file)
-        self.output_file = op.join(dirname, '%s  %s.m4a' % (self.artist, self.title))
+        stem = Path(mp4_file).stem
+        self.album, self.title, self.artist = stem.split('  ')
 
-        self.tempdir = tempfile.mkdtemp()
-        self.artwork_file = op.join(self.tempdir, 'artwork.jpg')
-        
-        if len(parts) > 3:
-            start, end = parts[3].split('-')
-            self.start = self.get_seconds(start)
-            self.end = self.get_seconds(end)
-        else:
-            self.start = self.end = None
+        parent = Path(mp4_file).parent
+        self.output_file = str(parent.joinpath(
+            '%s  %s.m4a' % (self.artist, self.title)))
 
-    def get_seconds(self, time_str):
-        min, sec = time_str.split(':')
-        return int(min) * 60 + int(sec)
+        self.artwork_file = None
+        for ext in ('.jpg', '.png'):
+            imgfile = parent.joinpath(stem + ext)
+            if imgfile.exists():
+                self.artwork_file = str(imgfile)
+                break
 
-    @property
-    def time_args(self):
-        "Return tuple with format (start, duration)."
-        if self.start is not None:
-            return (self.start, self.end - self.start)
-        else:
-            return None
+        if not self.artwork_file:
+            self.tempdir = tempfile.mkdtemp()
+            self.artwork_file = op.join(self.tempdir, 'artwork.jpg')
 
     def __str__(self):
-        return '%s - %s - %s (%s)' % (self.artist, self.title, self.album, 
-            self.time_args)
+        return '%s - %s - %s (%s)' % (self.artist, self.title, self.album)
 
     def __del__(self):
-        shutil.rmtree(self.tempdir)
+        if hasattr(self, 'tempdir') and op.exists(self.tempdir):
+            shutil.rmtree(self.tempdir)
 
 
 def extract_elements(song):
@@ -102,14 +91,15 @@ def extract_elements(song):
     ]
     subprocess.call(cmd)
 
-    cmd = [
-        'ffmpeg',
-        '-ss', '0',
-        '-i', song.input_file,
-        '-frames:v', '1',
-        song.artwork_file,
-    ]
-    subprocess.call(cmd)
+    if not op.exists(song.artwork_file):
+        cmd = [
+            'ffmpeg',
+            '-ss', '0',
+            '-i', song.input_file,
+            '-frames:v', '1',
+            song.artwork_file,
+        ]
+        subprocess.call(cmd)
 
 
 def adjust_gain(song):
@@ -120,22 +110,6 @@ def adjust_gain(song):
         song.output_file
     ]
     subprocess.call(cmd)
-
-
-def trim(song):
-    start, duration = song.time_args
-    tmpfile = op.join(song.tempdir, 'temp.m4a')
-
-    cmd = [
-        'ffmpeg',
-        '-ss', str(start),
-        '-i', song.output_file,
-        '-t', str(duration),
-        '-acodec', 'copy',
-        tmpfile,
-    ]
-    subprocess.call(cmd)
-    shutil.copy(tmpfile, song.output_file)
 
 
 def add_metadata(song):
@@ -161,11 +135,9 @@ if __name__ == '__main__':
     for mp4_file in mp4_files:
         try:
             song = Song(op.join(dirname, mp4_file))
-        except:
+        except Exception as ex:
             continue
 
         extract_elements(song)
         adjust_gain(song)
-        if song.time_args:
-            trim(song)
         add_metadata(song)
